Add button style in index.html & link to header
A specific button style css setting was added into a button that appears
in sign-up because it needs a white border that doesn't appear anywhere
else in the project. Links were added to the header instead of using a
pure list because it is more accurate to how a real landing page would
be ( even if they do not go anywhere ).